Re-organize query methods (follow-up of #21)
