+++ app/controllers/geo/Constants.java
@@ -1,0 +1,14 @@
+package controllers.geo;
+
+public class Constants {
+
+	public static final String TERM = "term";
+	public static final String STREET = "street";
+	public static final String CITY = "city";
+	public static final String COUNTRY = "country";
+	public static final String GEOCODE = "geocode";
+	public static final String POSTALCODE = "postalcode";
+	public static final String NOT_FOUND = "Not found (404): ";
+	public static final String SEPARATOR = ",";
+
+}
